eigsim/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "eigsim"
version = "0.1.0"
description = "Simulation code for the EIGSEP experiment"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "croissant-sim>=5.1",
    "numpy",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[tool.hatch.build.targets.wheel]
packages = ["src/eigsim"]
